Move deposit calculations
