+++ src/GraphQL/Schemas/Primary/Substrate/Mutations/SetCollectionAttributeMutation.php
@@ -3,10 +3,10 @@
 namespace Enjin\Platform\GraphQL\Schemas\Primary\Substrate\Mutations;
 
 use Closure;
-use Codec\Utils;
 use Enjin\Platform\GraphQL\Base\Mutation;
 use Enjin\Platform\GraphQL\Schemas\Primary\Substrate\Traits\InPrimarySubstrateSchema;
 use Enjin\Platform\GraphQL\Schemas\Primary\Traits\HasSkippableRules;
+use Enjin\Platform\GraphQL\Schemas\Primary\Traits\HasTransactionDeposit;
 use Enjin\Platform\GraphQL\Types\Input\Substrate\Traits\HasIdempotencyField;
 use Enjin\Platform\GraphQL\Types\Input\Substrate\Traits\HasSimulateField;
 use Enjin\Platform\Interfaces\PlatformBlockchainTransaction;
@@ -25,6 +25,7 @@ class SetCollectionAttributeMutation extends Mutation implements PlatformBlockch
     use HasIdempotencyField;
     use HasSkippableRules;
     use HasSimulateField;
+    use HasTransactionDeposit;
 
     /**
      * Get the mutation's attributes.
@@ -93,7 +94,7 @@ class SetCollectionAttributeMutation extends Mutation implements PlatformBlockch
                 'method' => $this->getMutationName(),
                 'encoded_data' => $encodedData,
                 'idempotency_key' => $args['idempotencyKey'] ?? Str::uuid()->toString(),
-                'deposit' => $this->getDepositValue($args),
+                'deposit' => $this->getSetCollectionAttributeDeposit($args),
                 'simulate' => $args['simulate'],
             ]),
             $resolveInfo
@@ -106,15 +107,6 @@ class SetCollectionAttributeMutation extends Mutation implements PlatformBlockch
     public function getMethodName(): string
     {
         return 'setAttribute';
-    }
-
-    protected function getDepositValue(array $args): ?string
-    {
-        $depositBase = gmp_init('200000000000000000');
-        $depositPerByte = gmp_init('100000000000000');
-        $totalBytes = count(Utils::string2ByteArray($args['key'])) + count(Utils::string2ByteArray($args['value']));
-
-        return gmp_strval(gmp_add($depositBase, gmp_mul($depositPerByte, $totalBytes)));
     }
 
     /**
